+++ sympy/physics/quantum/piab.py
@@ -5,7 +5,7 @@ from sympy import Symbol, pi, sqrt, sin, Interval, S
 from sympy.physics.quantum.operator import HermitianOperator
 from sympy.physics.quantum.state import Ket, Bra
 from sympy.physics.quantum.constants import hbar
-from sympy.physics.quantum.kronecker import KroneckerDelta
+from sympy.functions.special.tensor_functions import KroneckerDelta
 from sympy.physics.quantum.hilbert import L2
 
 m = Symbol('m')
